mcp/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.34,<0.35
httpx>=0.28,<0.29
pyyaml>=6.0,<6.1
pydantic>=2.10,<2.11
